requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-plugins==1.1.1
cligj==0.7.2
Fiona==1.9.4.post1
GeoAlchemy2==0.14.1
geopandas==0.13.2
idna==3.4
numpy==1.25.2
packaging==23.1
pandas==2.0.3
psycopg2==2.9.7
pyproj==3.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
shapely==2.0.1
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
